+++ dbt_platform_helper/utils/validation.py
@@ -1,15 +1,23 @@
 import ipaddress
 import re
+from pathlib import Path
 
 import click
+import yaml
 from botocore.exceptions import ClientError
 from schema import Optional
 from schema import Or
 from schema import Regex
 from schema import Schema
 from schema import SchemaError
+from yaml.parser import ParserError
 
+from dbt_platform_helper.constants import CODEBASE_PIPELINES_KEY
+from dbt_platform_helper.constants import ENVIRONMENTS_KEY
+from dbt_platform_helper.constants import PLATFORM_CONFIG_FILE
 from dbt_platform_helper.utils.aws import get_aws_session_or_abort
+from dbt_platform_helper.utils.files import apply_environment_defaults
+from dbt_platform_helper.utils.messages import abort_with_error
 
 
 def validate_string(regex_pattern: str):
@@ -228,6 +236,7 @@ POSTGRES_DEFINITION = {
             Optional("deletion_protection"): DELETION_PROTECTION,
             Optional("multi_az"): bool,
             Optional("storage_type"): POSTGRES_STORAGE_TYPES,
+            Optional("backup_retention_days"): int_between(1, 35),
         }
     },
     Optional("objects"): [
@@ -259,6 +268,12 @@ AURORA_DEFINITION = {
     ],
 }
 
+LIFECYCLE_RULE = {
+    Optional("filter_prefix"): str,
+    "expiration_days": int,
+    "enabled": bool,
+}
+
 S3_BASE = {
     Optional("readonly"): bool,
     Optional("services"): Or("__all__", [str]),
@@ -268,6 +283,7 @@ S3_BASE = {
             Optional("deletion_policy"): DELETION_POLICY,
             Optional("retention_policy"): RETENTION_POLICY,
             Optional("versioning"): bool,
+            Optional("lifecycle_rules"): [LIFECYCLE_RULE],
         }
     },
 }
@@ -415,6 +431,7 @@ CODEBASE_PIPELINES_DEFINITION = [
 
 ENVIRONMENT_PIPELINES_DEFINITION = {
     str: {
+        Optional("account"): str,
         Optional("branch", default="main"): str,
         "slack_channel": str,
         "trigger_on_push": bool,
@@ -446,6 +463,97 @@ PLATFORM_CONFIG_SCHEMA = Schema(
         Optional("environment_pipelines"): ENVIRONMENT_PIPELINES_DEFINITION,
     }
 )
+
+
+def validate_platform_config(config):
+    get_aws_session_or_abort()  # Ensure we have a valid session as validation requires it.
+    PLATFORM_CONFIG_SCHEMA.validate(config)
+    _validate_environment_pipelines(config)
+    _validate_codebase_pipelines(config)
+
+
+def _validate_environment_pipelines(config):
+    enriched_config = apply_environment_defaults(config)
+    bad_pipelines = {}
+
+    for pipeline_name, pipeline in enriched_config.get("environment_pipelines", {}).items():
+        bad_envs = []
+        pipeline_account = pipeline.get("account", None)
+        if pipeline_account:
+            for env in pipeline.get("environments", {}).keys():
+                env_account = (
+                    enriched_config.get("environments", {})
+                    .get(env, {})
+                    .get("accounts", {})
+                    .get("deploy", {})
+                    .get("name")
+                )
+                if not env_account == pipeline_account:
+                    bad_envs.append(env)
+        if bad_envs:
+            bad_pipelines[pipeline_name] = {"account": pipeline_account, "bad_envs": bad_envs}
+    if bad_pipelines:
+        message = "The following pipelines are misconfigured:"
+        for pipeline, detail in bad_pipelines.items():
+            envs = detail["bad_envs"]
+            acc = detail["account"]
+            message += f"  '{pipeline}' - these environments are not in the '{acc}' account: {', '.join(envs)}\n"
+        abort_with_error(message)
+
+
+def _validate_codebase_pipelines(pipeline_config):
+    if CODEBASE_PIPELINES_KEY in pipeline_config:
+        for codebase in pipeline_config[CODEBASE_PIPELINES_KEY]:
+            codebase_environments = []
+
+            for pipeline in codebase["pipelines"]:
+                codebase_environments += [e["name"] for e in pipeline[ENVIRONMENTS_KEY]]
+
+            unique_codebase_environments = sorted(list(set(codebase_environments)))
+
+            if sorted(codebase_environments) != sorted(unique_codebase_environments):
+                abort_with_error(
+                    f"The {PLATFORM_CONFIG_FILE} file is invalid, each environment can only be "
+                    "listed in a single pipeline per codebase"
+                )
+
+
+def load_and_validate_platform_config(path=PLATFORM_CONFIG_FILE):
+    config_file_check(path)
+    try:
+        conf = yaml.safe_load(Path(path).read_text())
+        validate_platform_config(conf)
+        return conf
+    except ParserError:
+        abort_with_error(f"{PLATFORM_CONFIG_FILE} is not valid YAML")
+
+
+def config_file_check(path=PLATFORM_CONFIG_FILE):
+    platform_config_exists = Path(path).exists()
+    errors = []
+
+    messages = {
+        "storage.yml": " under the key 'extensions'",
+        "extensions.yml": " under the key 'extensions'",
+        "pipelines.yml": ", change the key 'codebases' to 'codebase_pipelines'",
+    }
+
+    for file in messages.keys():
+        if Path(file).exists():
+            if platform_config_exists:
+                message = f"`{file}` has been superseded by `{PLATFORM_CONFIG_FILE}` and should be deleted."
+            else:
+                message = f"`{file}` is no longer supported. Please move its contents into a file named `{PLATFORM_CONFIG_FILE}`{messages[file]} and delete `{file}`."
+            errors.append(message)
+
+    if not errors and not platform_config_exists:
+        errors.append(
+            f"`{PLATFORM_CONFIG_FILE}` is missing. Please check it exists and you are in the root directory of your deployment project."
+        )
+
+    if errors:
+        click.secho("\n".join(errors), bg="red")
+        exit(1)
 
 
 S3_SCHEMA = Schema(S3_DEFINITION)