+++ platform_helper.py
@@ -12,6 +12,7 @@ from dbt_platform_helper.commands.codebase import codebase as codebase_commands
 from dbt_platform_helper.commands.conduit import conduit as conduit_commands
 from dbt_platform_helper.commands.config import config as config_commands
 from dbt_platform_helper.commands.copilot import copilot as copilot_commands
+from dbt_platform_helper.commands.database import database as database_commands
 from dbt_platform_helper.commands.dns import cdn as cdn_commands
 from dbt_platform_helper.commands.dns import domain as domain_commands
 from dbt_platform_helper.commands.environment import environment as environment_commands
@@ -44,6 +45,7 @@ platform_helper.add_command(generate_commands)
 platform_helper.add_command(pipeline_commands)
 platform_helper.add_command(secrets_commands)
 platform_helper.add_command(notify_commands)
+platform_helper.add_command(database_commands)
 
 if __name__ == "__main__":
     platform_helper()