+++ regression_tests/stages/run_demodjango_smoke_tests.sh
@@ -8,4 +8,8 @@ echo -e "\nInstall dependencies"
 poetry install
 
 echo -e "\nRun smoke tests"
+<<<<<<< HEAD
 ./smoke_tests.sh ${TARGET_ENVIRONMENT}
+=======
+./tests/browser/run.sh toolspr smoke
+>>>>>>> main